✨ feat: page는 server 컴포넌트는 client로 변경
